+++ image-builder-backend/api-gateway-service/.gitignore
@@ -1,1 +1,1 @@
-/api
+*.gen.go
