+++ image-builder-backend/api-gateway-service/Makefile
@@ -1,8 +1,17 @@
+OPENAPI_FILE = ../../openapi.yml
+TARGET_DIR = ./cmd/api
+
 .PHONY: generate
 
 all: generate
 
 generate:
-	mkdir api
-	oapi-codegen -package api -generate types,gin-server ../../openapi.yml > ./api/api.gen.go
+	# generate types and server into different files
+	# for better separation of concerns
+	oapi-codegen -generate types \
+		-package api \
+		$(OPENAPI_FILE) > $(TARGET_DIR)/types.gen.go
+	oapi-codegen -generate gin-server \
+		-package api \
+		$(OPENAPI_FILE) > $(TARGET_DIR)/interface.gen.go
 	go mod tidy
