+++ image-builder-backend/api-gateway-service/README.md
@@ -1,8 +1,7 @@
 # ApiGatewayService
 
-
-
 ## How to install and run backend
+install oapi-codegen?
 
 ### generate api from openapi file
 just run
