+++ image-builder-backend/api-gateway-service/cmd/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/api-gateway-service/cmd/api"
+)
+
+func main() {
+	api.StartServer()
+}
